move all http/send test to own module
